[Rename] pages/User, components/User -> pages/Auth, components/Auth
