+++ src/pages/Auth/DeleteAccount.jsx
