+++ src/pages/Auth/SignIn.jsx
@@ -5,10 +5,10 @@ import { handleError } from '../../utils/common';
 import { toast } from 'react-toastify';
 import Logo from '../../components/Global/Logo';
 import BackButton from '../../components/Global/BackButton';
-import SocialLoginButton from '../../components/User/SignIn/SocialLoginButton';
-import InputField from '../../components/User/SignIn/InputField';
-import ErrorMessage from '../../components/User/SignIn/ErrorMessage';
-import TextLink from '../../components/User/SignIn/TextLink';
+import SocialLoginButton from '../../components/Auth/SignIn/SocialLoginButton';
+import InputField from '../../components/Auth/SignIn/InputField';
+import ErrorMessage from '../../components/Auth/SignIn/ErrorMessage';
+import TextLink from '../../components/Auth/SignIn/TextLink';
 import LOGO_GOOGLE from '/assets/img/logo_google.webp';
 import LOGO_KAKAO from '/assets/img/logo_kakao.webp';
 import LOGO_NAVER from '/assets/img/logo_naver.webp';
